requirements-ai-proxy.txt
# PowerSave AI Proxy Server Dependencies

# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For production deployment
gunicorn>=21.2.0
